Relocate worklock with the rest of the cli commands
